+++ packages/stream_chat_flutter/example/lib/tutorial_part_3.dart
@@ -81,7 +81,7 @@ class ChannelListPage extends StatelessWidget {
         child: ChannelListView(
           filter: Filter.in_(
             'members',
-            [StreamChat.of(context).user!.id],
+            [StreamChat.of(context).currentUser!.id],
           ),
           channelPreviewBuilder: _channelPreviewBuilder,
           // sort: [SortOption('last_message_at')],