+++ packages/stream_chat_flutter/lib/src/channel_list_view.dart
@@ -553,7 +553,7 @@ class _ChannelListViewState extends State<ChannelListView> {
                 'owner',
               ].contains(channel.state!.members
                   .firstWhereOrNull(
-                      (m) => m.userId == channel.client.state.user?.id)
+                      (m) => m.userId == channel.client.state.currentUser?.id)
                   ?.role))
                 IconSlideAction(
                   color: backgroundColor,