+++ packages/stream_chat_flutter/lib/src/message_widget.dart
@@ -92,6 +92,8 @@ class MessageWidget extends StatefulWidget {
     this.userAvatarBuilder,
     this.editMessageInputBuilder,
     this.textBuilder,
+    this.bottomRowBuilder,
+    this.deletedBottomRowBuilder,
     this.onReturnAction,
     Map<String, AttachmentBuilder>? customAttachmentBuilders,
     this.readList,
@@ -275,6 +277,12 @@ class MessageWidget extends StatefulWidget {
   /// Function called on long press
   final void Function(BuildContext, Message)? onMessageActions;
 
+  /// Widget builder for building a bottom row below the message
+  final Widget Function(BuildContext, Message)? bottomRowBuilder;
+
+  /// Widget builder for building a bottom row below a deleted message
+  final Widget Function(BuildContext, Message)? deletedBottomRowBuilder;
+
   /// Widget builder for building user avatar
   final Widget Function(BuildContext, User)? userAvatarBuilder;
 
@@ -410,6 +418,8 @@ class MessageWidget extends StatefulWidget {
     Widget Function(BuildContext, Message)? editMessageInputBuilder,
     Widget Function(BuildContext, Message)? textBuilder,
     Widget Function(BuildContext, Message)? usernameBuilder,
+    Widget Function(BuildContext, Message)? bottomRowBuilder,
+    Widget Function(BuildContext, Message)? deletedBottomRowBuilder,
     void Function(BuildContext, Message)? onMessageActions,
     Message? message,
     MessageTheme? messageTheme,
@@ -463,6 +473,9 @@ class MessageWidget extends StatefulWidget {
             editMessageInputBuilder ?? this.editMessageInputBuilder,
         textBuilder: textBuilder ?? this.textBuilder,
         usernameBuilder: usernameBuilder ?? this.usernameBuilder,
+        bottomRowBuilder: bottomRowBuilder ?? this.bottomRowBuilder,
+        deletedBottomRowBuilder:
+            deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
         onMessageActions: onMessageActions ?? this.onMessageActions,
         message: message ?? this.message,
         messageTheme: messageTheme ?? this.messageTheme,
@@ -782,7 +795,11 @@ class _MessageWidgetState extends State<MessageWidget>
                             bottom:
                                 isPinned && widget.showPinHighlight ? 6.0 : 0.0,
                           ),
-                          child: _bottomRow,
+                          child: widget.bottomRowBuilder?.call(
+                                context,
+                                widget.message,
+                              ) ??
+                              _bottomRow,
                         ),
                       if (isFailedState)
                         Positioned(
@@ -810,7 +827,7 @@ class _MessageWidgetState extends State<MessageWidget>
   }
 
   Widget _buildQuotedMessage() {
-    final isMyMessage = widget.message.user?.id == _streamChat.user?.id;
+    final isMyMessage = widget.message.user?.id == _streamChat.currentUser?.id;
     final onTap = widget.message.quotedMessage?.isDeleted != true &&
             widget.onQuotedMessageTap != null
         ? () => widget.onQuotedMessageTap!(widget.message.quotedMessageId)
@@ -830,22 +847,11 @@ class _MessageWidgetState extends State<MessageWidget>
 
   Widget get _bottomRow {
     if (isDeleted) {
-      final chatThemeData = _streamChatTheme;
-      return Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          StreamSvgIcon.eye(
-            color: chatThemeData.colorTheme.textLowEmphasis,
-            size: 16,
-          ),
-          const SizedBox(width: 8),
-          Text(
-            context.translations.onlyVisibleToYouText,
-            style: chatThemeData.textTheme.footnote
-                .copyWith(color: chatThemeData.colorTheme.textLowEmphasis),
-          ),
-        ],
-      );
+      return widget.deletedBottomRowBuilder?.call(
+            context,
+            widget.message,
+          ) ??
+          const Offstage();
     }
 
     final children = <Widget>[];
@@ -993,7 +999,7 @@ class _MessageWidgetState extends State<MessageWidget>
   Widget _buildReactionIndicator(
     BuildContext context,
   ) {
-    final ownId = _streamChat.user!.id;
+    final ownId = _streamChat.currentUser!.id;
     final reactionsMap = <String, Reaction>{};
     widget.message.latestReactions?.forEach((element) {
       if (!reactionsMap.containsKey(element.type) ||
@@ -1054,10 +1060,10 @@ class _MessageWidgetState extends State<MessageWidget>
                   showReactionPickerIndicator: widget.showReactions &&
                       (widget.message.status == MessageSendingStatus.sent),
                   showPinHighlight: false,
-                  showUserAvatar:
-                      widget.message.user!.id == channel.client.state.user!.id
-                          ? DisplayWidget.gone
-                          : DisplayWidget.show,
+                  showUserAvatar: widget.message.user!.id ==
+                          channel.client.state.currentUser!.id
+                      ? DisplayWidget.gone
+                      : DisplayWidget.show,
                 ),
                 onCopyTap: (message) =>
                     Clipboard.setData(ClipboardData(text: message.text)),
@@ -1118,7 +1124,7 @@ class _MessageWidgetState extends State<MessageWidget>
                 (widget.message.status == MessageSendingStatus.sent),
             showPinHighlight: false,
             showUserAvatar:
-                widget.message.user!.id == channel.client.state.user!.id
+                widget.message.user!.id == channel.client.state.currentUser!.id
                     ? DisplayWidget.gone
                     : DisplayWidget.show,
           ),
@@ -1279,7 +1285,7 @@ class _MessageWidgetState extends State<MessageWidget>
 
   Widget _buildPinnedMessage(Message message) {
     final pinnedBy = message.pinnedBy!;
-    final currentUser = _streamChat.user!;
+    final currentUser = _streamChat.currentUser!;
 
     return Padding(
       padding: const EdgeInsets.only(left: 8, right: 8, top: 4, bottom: 8),