+++ packages/stream_chat/CHANGELOG.md
@@ -1,8 +1,14 @@
 ## Upcoming
 
+🔄 Changed
+
+- `client.user` is now deprecated in favor of `client.currentUser`.
+- `client.userStream` is now deprecated in favor of `client.currentUserStream`.
+
 🐞 Fixed
 
 - [#563](https://github.com/GetStream/stream-chat-flutter/issues/563): `Channel.stopWatching()` not working 
+- [#575](https://github.com/GetStream/stream-chat-flutter/issues/575): Wrong `OwnUser.*`
 
 ## 2.0.0