+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,7 +1,14 @@
 ## Upcoming
 
+🛑️ Breaking Changes from `2.0.0`
+- Changed default message filter of `MessageListCore`
+
 ✅ Added
 - Added `MessageListCore.paginationLimit`
+
+🔄 Changed
+- `StreamChatCore.of(context).user` is now deprecated in favor of `StreamChatCore.of(context).currentUser`.
+- `StreamChatCore.of(context).userStream` is now deprecated in favor of `StreamChatCore.of(context).currentUserStream`.
 
 ## 2.0.0